[sil-ownership] Rename -enable-sil-ownership => -enable-sil-ownership-verifier
This option no longer controls whether or not ownership is enabled. Instead we
always emit in SILGen with ownership enabled and just conditionally enable the
verifier.

In terms of SIL parsing, as mentioned in previous commits, we now control
whether or not we are parsing sil with ownership via the
-assume-parsing-unqualified-ownership-sil. In the future, I believe that option
will go away as well once SIL ownership stripping moves out of the mandatory
pipeline. But that is another patch for another day.