Disable overlay resilience in the 4.2 release.
The 4.2 release is not ABI-stable, so it's not strictly necessary to enable resilience.
We did so in order to gain information and find unanticipated bugs ahead of ABI stability.
Now that a number of such bugs have been discovered, it makes sense to disable resilience
in the release in order to reduce the impact on users.

rdar://42756146