Removed code and cleaned up tests related to multi-result tensors encoded as a tuple (#18394)
Removed code and cleaned up tests related to encoding multi-result tensors,
which should never occur in the world of graph_opt insts, which are multi-valued
instructions. Added a unit test on tuples of tensor values.

We eagerly removed such code to simplify the partition code base, making it
easier to make other enhancements and bug fixes.