disable send/recv warnings for scalars (#18549)
Based on a discussion with @mhong, this disables send/recv warnings for values that are definitely scalar.

This cleans up the warning output (e.g. removes warnings about sends/recvs for outer loop indices), making it easier to focus on more interesting warnings.

At some point we might want to re-enable some scalar warnings based on some cost model (e.g. warn for scalar transfers in inner loops), but that's a lot more complicated and this is a good starting point.