master-next: Disable LLVM ABI breaking check enforcement in the demangler
LLVM r338955 added an include of "llvm/Config/abi-breaking.h" in
the llvm/ADT/STLExtras.h header so that the Swift demangler's use of
header-only ADT code runs afoul of the ABI breaking check enforcement.
Disable those checks for the demangler in the same way that we already
do for the Swift runtime.