ClangImporter: Convert function/method/accessor import to use ParameterList::getParams()
Note that these functions used to return a FunctionType; now, they
return the result type instead:

- importFunctionType()
- importMethodType()
- importAccessorMethodType()

The parameters were already being returned in the form of a ParameterList
of ParamDecls.

This actually has the effect of making the code simpler, because callers
would always take them apart and build new function types anyway.