Per discussion with @lattner, adjusted this patch to only do ssa value propagation on SIL insts related to tfop operands (including attrs).
To make sure that we can propagate values of $TensorShape struct type:
1. In `explodeAggregateInst()`, changed `emitLoweredXXXMostDerivedDescendents()`
calls to `emitLoweredXXXDirectChildren()`, so that we do not dig into and
explode structs that are unrelated to tensorflow value types. One such example
is $Array<Int32>, which is a struct field of $TensorShape.

2. Extended SROA to allow skipping the explosion of alloc_stack over the
$TensorShape type.

While we may need to further enhance deabstraction code to propagate through
more SIL patterns (especially related to const array handling), the code is
structured in a way that's reasonably easy to extend.