[analysis] Turn on the pass manager notification verifier analysis.
When asserts are enabled this tracks the add/delete notifications from the
SILPassManager it belongs to. When the SILPassManager is destroyed, the analysis
verifies that it worked correctly by looping over all the functions in the
module and verifying that its internal state is consistent with the module.

This will ensure we never lose add/delete notification events ever again by
mistake.

rdar://42301529