[Type Checker] Add TypeResolution(Stage) to describe type computations.
Introduce an enum to describe the “stage” of type resolution that should
be performed, which is a notion that loosely exists within the type
checker already. Encapsulate the notion of a stage + the DeclContext
in which a type-check occurs into a new TypeResolution class, which
will replace the current GenericTypeResolver hierarchy.