[stdlib] Update Policy.swift to use designated protocols for operators.
Also enable the parsing of these with
-enable-operator-designated-protocols in the stdlib build.

The constraint solver support to make use of these is not yet in-tree.
and will probably land with a different command-line option to enable
it, so that we can continue to have parsing support enabled in-tree
while the constraint solver support is iteratively improved.