Sema: Tweak 'penalize conversions to Any' hack
We would previously skip this code path when calling a function
with a single 'Any' argument. I believe it was totally not
intentional, however it does mean that fewer solutions were
getting this penalty than you would expect.

Restore the old behavior by not penalizing the solution when
performing an argument conversion to 'Any'. However, keep the
penalty when converting functions to 'Any', to decrease the
odds that we'll pass a trailing closure as an 'Any'.