Deprecate unsafe[Adding,Subtracting,Multiplied,Divided].
AFAICT these functions are essentially unused, and their semantics
are quite confusing. People tend to believe that they are equivalent
to the corresponding &-prefixed operators, but they absolutely are
not, and the way in which they differ is quite dangerous.

These functions trap if they overflow in debug, and in optimized
builds, overflow is *undefined* rather than wrapping. This is a huge
footgun, which might be justifiable if they were being used for
performance purposes, but they are not.